Update Gesture_Controller.py
For some people pinch gesture is not working properly for changing the brightness, it may be dur to the fade_brightness function which is not working and now i changed to it set brightness function and it works properly now. Along with that this code is working only for the right hand i made a small correction to make it work for both the hands so that which hand is get detected that hand can be used to operate the system. Now the code works fine in everything it was written to :)